java-api-client v1.0.0

+++ src/main/java/dev/astatic/nodestyclient/api/models/dedicated/DedicatedServerHardwareComponent.java
@@ -3,7 +3,6 @@ package dev.astatic.nodestyclient.api.models.dedicated;
 public record DedicatedServerHardwareComponent(
         String component,
         String model,
-        double value,
+        int value,
         String valueSuffix
-) {
-}
+) {}
